+++ go/integTest/shared_commands_test.go
@@ -4703,3 +4703,268 @@ func (suite *GlideTestSuite) TestPersist() {
 		assert.False(t, resultInvalidKey.Value())
 	})
 }
+
+func (suite *GlideTestSuite) TestSortWithOptions_ExternalWeights() {
+	suite.SkipIfServerVersionLowerThanBy("8.0.0")
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key := uuid.New().String()
+		client.LPush(key, []string{"item1", "item2", "item3"})
+
+		client.Set("weight_item1", "3")
+		client.Set("weight_item2", "1")
+		client.Set("weight_item3", "2")
+
+		options := options.NewSortOptions().
+			SetByPattern("weight_*").
+			SetOrderBy(options.ASC).
+			SetIsAlpha(false)
+
+		sortResult, err := client.SortWithOptions(key, options)
+
+		assert.Nil(suite.T(), err)
+		resultList := []api.Result[string]{
+			api.CreateStringResult("item2"),
+			api.CreateStringResult("item3"),
+			api.CreateStringResult("item1"),
+		}
+
+		assert.Equal(suite.T(), resultList, sortResult)
+	})
+}
+
+func (suite *GlideTestSuite) TestSortWithOptions_GetPatterns() {
+	suite.SkipIfServerVersionLowerThanBy("8.0.0")
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key := uuid.New().String()
+		client.LPush(key, []string{"item1", "item2", "item3"})
+
+		client.Set("object_item1", "Object_1")
+		client.Set("object_item2", "Object_2")
+		client.Set("object_item3", "Object_3")
+
+		options := options.NewSortOptions().
+			SetByPattern("weight_*").
+			SetOrderBy(options.ASC).
+			SetIsAlpha(false).
+			AddGetPattern("object_*")
+
+		sortResult, err := client.SortWithOptions(key, options)
+
+		assert.Nil(suite.T(), err)
+
+		resultList := []api.Result[string]{
+			api.CreateStringResult("Object_2"),
+			api.CreateStringResult("Object_3"),
+			api.CreateStringResult("Object_1"),
+		}
+
+		assert.Equal(suite.T(), resultList, sortResult)
+	})
+}
+
+func (suite *GlideTestSuite) TestSortWithOptions_SuccessfulSortByWeightAndGet() {
+	suite.SkipIfServerVersionLowerThanBy("8.0.0")
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key := uuid.New().String()
+		client.LPush(key, []string{"item1", "item2", "item3"})
+
+		client.Set("weight_item1", "10")
+		client.Set("weight_item2", "5")
+		client.Set("weight_item3", "15")
+
+		client.Set("object_item1", "Object 1")
+		client.Set("object_item2", "Object 2")
+		client.Set("object_item3", "Object 3")
+
+		options := options.NewSortOptions().
+			SetOrderBy(options.ASC).
+			SetIsAlpha(false).
+			SetByPattern("weight_*").
+			AddGetPattern("object_*").
+			AddGetPattern("#")
+
+		sortResult, err := client.SortWithOptions(key, options)
+
+		assert.Nil(suite.T(), err)
+
+		resultList := []api.Result[string]{
+			api.CreateStringResult("Object 2"),
+			api.CreateStringResult("item2"),
+			api.CreateStringResult("Object 1"),
+			api.CreateStringResult("item1"),
+			api.CreateStringResult("Object 3"),
+			api.CreateStringResult("item3"),
+		}
+
+		assert.Equal(suite.T(), resultList, sortResult)
+
+		objectItem2, err := client.Get("object_item2")
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "Object 2", objectItem2.Value())
+
+		objectItem1, err := client.Get("object_item1")
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "Object 1", objectItem1.Value())
+
+		objectItem3, err := client.Get("object_item3")
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "Object 3", objectItem3.Value())
+
+		assert.Equal(suite.T(), "item2", sortResult[1].Value())
+		assert.Equal(suite.T(), "item1", sortResult[3].Value())
+		assert.Equal(suite.T(), "item3", sortResult[5].Value())
+	})
+}
+
+func (suite *GlideTestSuite) TestSortStoreWithOptions_ByPattern() {
+	suite.SkipIfServerVersionLowerThanBy("8.0.0")
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key := "{listKey}" + uuid.New().String()
+		sortedKey := "{listKey}" + uuid.New().String()
+		client.LPush(key, []string{"a", "b", "c", "d", "e"})
+		client.Set("{listKey}weight_a", "5")
+		client.Set("{listKey}weight_b", "2")
+		client.Set("{listKey}weight_c", "3")
+		client.Set("{listKey}weight_d", "1")
+		client.Set("{listKey}weight_e", "4")
+
+		options := options.NewSortOptions().SetByPattern("{listKey}weight_*")
+
+		result, err := client.SortStoreWithOptions(key, sortedKey, options)
+
+		assert.Nil(suite.T(), err)
+		assert.NotNil(suite.T(), result)
+		assert.Equal(suite.T(), int64(5), result.Value())
+
+		sortedValues, err := client.LRange(sortedKey, 0, -1)
+		resultList := []api.Result[string]{
+			api.CreateStringResult("d"),
+			api.CreateStringResult("b"),
+			api.CreateStringResult("c"),
+			api.CreateStringResult("e"),
+			api.CreateStringResult("a"),
+		}
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), resultList, sortedValues)
+	})
+}
+
+func (suite *GlideTestSuite) TestSortReadOnlyWithOptions_ExternalWeights() {
+	suite.SkipIfServerVersionLowerThanBy("8.0.0")
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		if suite.serverVersion < "7.0.0" {
+			suite.T().Skip("This feature is added in version 7")
+		}
+		key := uuid.New().String()
+		client.LPush(key, []string{"item1", "item2", "item3"})
+
+		client.Set("weight_item1", "3")
+		client.Set("weight_item2", "1")
+		client.Set("weight_item3", "2")
+
+		options := options.NewSortOptions().
+			SetByPattern("weight_*").
+			SetOrderBy(options.ASC).
+			SetIsAlpha(false)
+
+		sortResult, err := client.SortReadOnlyWithOptions(key, options)
+
+		assert.Nil(suite.T(), err)
+		resultList := []api.Result[string]{
+			api.CreateStringResult("item2"),
+			api.CreateStringResult("item3"),
+			api.CreateStringResult("item1"),
+		}
+		assert.Equal(suite.T(), resultList, sortResult)
+	})
+}
+
+func (suite *GlideTestSuite) TestSortReadOnlyWithOptions_GetPatterns() {
+	suite.SkipIfServerVersionLowerThanBy("8.0.0")
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		if suite.serverVersion < "7.0.0" {
+			suite.T().Skip("This feature is added in version 7")
+		}
+		key := uuid.New().String()
+		client.LPush(key, []string{"item1", "item2", "item3"})
+
+		client.Set("object_item1", "Object_1")
+		client.Set("object_item2", "Object_2")
+		client.Set("object_item3", "Object_3")
+
+		options := options.NewSortOptions().
+			SetByPattern("weight_*").
+			SetOrderBy(options.ASC).
+			SetIsAlpha(false).
+			AddGetPattern("object_*")
+
+		sortResult, err := client.SortReadOnlyWithOptions(key, options)
+
+		assert.Nil(suite.T(), err)
+
+		resultList := []api.Result[string]{
+			api.CreateStringResult("Object_2"),
+			api.CreateStringResult("Object_3"),
+			api.CreateStringResult("Object_1"),
+		}
+
+		assert.Equal(suite.T(), resultList, sortResult)
+	})
+}
+
+func (suite *GlideTestSuite) TestSortReadOnlyWithOptions_SuccessfulSortByWeightAndGet() {
+	suite.SkipIfServerVersionLowerThanBy("8.0.0")
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		if suite.serverVersion < "7.0.0" {
+			suite.T().Skip("This feature is added in version 7")
+		}
+		key := uuid.New().String()
+		client.LPush(key, []string{"item1", "item2", "item3"})
+
+		client.Set("weight_item1", "10")
+		client.Set("weight_item2", "5")
+		client.Set("weight_item3", "15")
+
+		client.Set("object_item1", "Object 1")
+		client.Set("object_item2", "Object 2")
+		client.Set("object_item3", "Object 3")
+
+		options := options.NewSortOptions().
+			SetOrderBy(options.ASC).
+			SetIsAlpha(false).
+			SetByPattern("weight_*").
+			AddGetPattern("object_*").
+			AddGetPattern("#")
+
+		sortResult, err := client.SortReadOnlyWithOptions(key, options)
+
+		assert.Nil(suite.T(), err)
+
+		resultList := []api.Result[string]{
+			api.CreateStringResult("Object 2"),
+			api.CreateStringResult("item2"),
+			api.CreateStringResult("Object 1"),
+			api.CreateStringResult("item1"),
+			api.CreateStringResult("Object 3"),
+			api.CreateStringResult("item3"),
+		}
+
+		assert.Equal(suite.T(), resultList, sortResult)
+
+		objectItem2, err := client.Get("object_item2")
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "Object 2", objectItem2.Value())
+
+		objectItem1, err := client.Get("object_item1")
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "Object 1", objectItem1.Value())
+
+		objectItem3, err := client.Get("object_item3")
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "Object 3", objectItem3.Value())
+
+		assert.Equal(suite.T(), "item2", sortResult[1].Value())
+		assert.Equal(suite.T(), "item1", sortResult[3].Value())
+		assert.Equal(suite.T(), "item3", sortResult[5].Value())
+	})
+}
